# ENH:  Use SVML for fp32 and fp64 power and arctan2