--- a/numpy/core/tests/test_umath.py
+++ b/numpy/core/tests/test_umath.py
@@ -1098,10 +1098,11 @@ class TestPower:
     
     def test_float_to_inf_power(self):
         for dt in [np.float32, np.float64]:
-            a = np.array([1,1,2,2,-2,-2,np.inf,-np.inf], dt)
-            b = np.array([np.inf, -np.inf, np.inf, -np.inf, np.inf, -np.inf, np.inf, -np.inf], dt)
-            r = np.array([1,1,np.inf,0,np.inf,0,np.inf,0], dt)
-            assert_equal(np.power(a,b), r)
+            a = np.array([1, 1, 2, 2, -2, -2, np.inf, -np.inf], dt)
+            b = np.array([np.inf, -np.inf, np.inf, -np.inf,
+                                np.inf, -np.inf, np.inf, -np.inf], dt)
+            r = np.array([1, 1, np.inf, 0, np.inf, 0, np.inf, 0], dt)
+            assert_equal(np.power(a, b), r)
 
 
 class TestFloat_power: