… numpy/core/code_generators/generate_umath.py
… numpy/core/src/common/npy_svml.h