<!--         ----------------------------------------------------------------
                MAKE SURE YOUR PR GETS THE ATTENTION IT DESERVES!
                ----------------------------------------------------------------

*  FORMAT IT RIGHT:
      https://www.numpy.org/devdocs/dev/development_workflow.html#writing-the-commit-message

*  IF IT'S A NEW FEATURE OR API CHANGE, TEST THE WATERS:
      https://www.numpy.org/devdocs/dev/development_workflow.html#get-the-mailing-list-s-opinion

*  HIT ALL THE GUIDELINES:
      https://numpy.org/devdocs/dev/index.html#guidelines

*  WHAT TO DO IF WE HAVEN'T GOTTEN BACK TO YOU:
      https://www.numpy.org/devdocs/dev/development_workflow.html#getting-your-pr-reviewed
-->
This adds support for SVML's power and arctan2 for 32 and 64 bit floating point. On a simple benchmark of these functions included in this merge request, this gives a very significant performance improvement. It also adds some tests for power's behavior on infinities and NANs.
```
       before           after         ratio
     [397b6e51]       [c85e17e3]
     <main>           <SVML>
-     14.5±0.05ms      2.29±0.07ms     0.16  bench_ufunc.BinaryBench.time_pow_64
-     6.12±0.02ms         954±10μs     0.16  bench_ufunc.BinaryBench.time_pow_32
-      16.2±0.2ms      1.93±0.04ms     0.12  bench_ufunc.BinaryBench.time_atan2_64
-     18.3±0.06ms          822±8μs     0.04  bench_ufunc.BinaryBench.time_atan2_32

```